requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database - MongoDB Only
pymongo>=4.5.0

# Authentication & JWT
PyJWT>=2.8.0
cryptography>=41.0.0

# Environment & configuration
python-dotenv>=1.0.0

# AWS/S3
boto3>=1.26.0
botocore>=1.29.0

# Supabase
supabase>=2.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Password hashing
bcrypt>=4.0.0

# OAuth & Social Media
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
aiohttp>=3.8.0

# Additional dependencies for Heroku
gunicorn>=21.2.0
passlib[bcrypt]>=1.7.4

# SQLAlchemy (required by some modules)
sqlalchemy>=2.0.0

# Additional dependencies that might be missing
psycopg2-binary>=2.9.0
email-validator>=2.0.0

# Additional dependencies for production
python-multipart>=0.0.6
jinja2>=3.1.0
